+++ src/core/index.ts
@@ -29,3 +29,6 @@ export const store = storeImport;
 
 import * as utilitiesImport from "./utilities";
 export const utilities = utilitiesImport;
+
+import * as identifiersImport from "./identifiers";
+export const identifiers = identifiersImport;